Refactor base1.py to make space
